Move `LogLevel` class to utilities
